+++ Podfile
@@ -114,6 +114,9 @@ target 'p2p_wallet' do
 
   # Sentry
   pod 'Sentry', :git => 'https://github.com/getsentry/sentry-cocoa.git', :tag => '7.31.5'
+  
+  # Amplitude
+  pod 'Amplitude', '~> 8.3.0'
 
   # AppsFlyer
   pod 'AppsFlyerFramework'
