+++ p2p_wallet/Analytics/Providers/ScreenAnalyticTracker/AnalyticView.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol AnalyticView {
+    var analyticId: String { get }
+}
