+++ p2p_wallet/Analytics/Providers/ScreenAnalyticTracker/ScreenAnalyticId.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol ScreenAnalyticId {
+    var analyticId: String { get }
+}
