+++ p2p_wallet/Analytics/Providers/ScreenAnalyticTracker/ScreenAnalyticTracker.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public class ScreenAnalyticTracker {
+    public static let shared = ScreenAnalyticTracker()
+
+    private(set) var currentViewId = ""
+
+    private init() {}
+
+    public func setCurrentViewId(_ viewId: String) {
+        currentViewId = viewId
+    }
+}
