--- p2p_wallet/Analytics/ScreenAnalyticTracker/AnalyticView.swift
@@ -1,12 +1,0 @@
-//
-//  AnalyticView.swift
-//  p2p_wallet
-//
-//  Created by Ivan on 14.09.2022.
-//
-
-import Foundation
-
-protocol AnalyticView {
-    var analyticId: String { get }
-}
