--- p2p_wallet/Analytics/ScreenAnalyticTracker/ScreenAnalyticId.swift
@@ -1,12 +1,0 @@
-//
-//  ScreenAnalyticId.swift
-//  p2p_wallet
-//
-//  Created by Ivan on 14.09.2022.
-//
-
-import Foundation
-
-protocol ScreenAnalyticId {
-    var analyticId: String { get }
-}
