--- p2p_wallet/Analytics/ScreenAnalyticTracker/ScreenAnalyticTracker.swift
@@ -1,20 +1,0 @@
-//
-//  ScreenAnalyticTracker.swift
-//  p2p_wallet
-//
-//  Created by Ivan on 14.09.2022.
-//
-
-import Foundation
-
-class ScreenAnalyticTracker {
-    static let shared = ScreenAnalyticTracker()
-
-    private(set) var currentViewId = ""
-
-    private init() {}
-
-    func setCurrentViewId(_ viewId: String) {
-        currentViewId = viewId
-    }
-}
