+++ p2p_wallet/Extensions/AnalyticsManager/AnalyticsManager+Extensions.swift
