+++ p2p_wallet/Extensions/AnalyticsManager/MirrorableEnum.swift
